Update tracker_parser.py
OGNTRK status messages may sometimes not conform to the expected format (new data for name of pilot is missing a load of fields). This caused the regex to fail, the match was not assigned, then tried to call a function on a null pointer. Try/except added to throw a ParseError which is dealt with by the calling code.